requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
astroid==2.5.6
attrs==20.3.0
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pylint==2.8.3
pyparsing==2.4.7
pytest==6.2.3
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
requests==2.22.0
jsonpath
json